server/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
mcp[cli]
requests
google-genai
python-dotenv
